Move SootherOptions into smoothers module
