+++ py_simplemg/__init__.py
@@ -1,5 +1,6 @@
 
 from __future__ import absolute_import
-from .multigrid_base import MultigridOptions, SmootherOptions
+from .multigrid_base import MultigridOptions
 from .multigrid_solver import solve_multigrid
+from .smoothers import SmootherOptions
